requirement.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-environ==0.12.0
django-filter==24.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==21.2.0
idna==3.10
pillow==11.1.0
psycopg==3.2.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
redis==5.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.27.1
social-auth-app-django==5.5.1
django-ckeditor==6.7.3
drf-spectacular==0.27.2